+++ docs/.vuepress/components/ValidationExample.vue
@@ -9,22 +9,32 @@
       <p v-if="form.samePassword.$dirty && form.samePassword.match.$invalid">
         {{ form.samePassword.match.$message }}
       </p>
+
+      <br />
+      <input
+        type="submit"
+        v-model="submitText"
+        :class="{
+          invalid: form.$anyDirty && form.$anyInvalid,
+          dirty: form.$anyDirty && !form.$anyInvalid,
+          error: form.$errors.length > 0
+        }"
+      />
     </form>
-    {{ form }}
   </div>
 </template>
 
 <script>
-import { createComponent, ref, reactive } from "@vue/composition-api";
+import { createComponent, ref, reactive, computed } from "@vue/composition-api";
 import { useValidation } from "vue-composable";
 
 const required = x => !!x;
 
 export default createComponent({
   setup() {
-    const name = ref("pikax");
-    const surname = ref("stuff");
-    const password = ref("123456");
+    const name = ref("");
+    const surname = ref("");
+    const password = ref("");
 
     const form = useValidation({
       firstName: {
@@ -54,12 +64,33 @@ export default createComponent({
       }
     });
 
-    const onSubmit = () => {
-      alert("submit form");
+    const submitText = computed(() => {
+      if (form.$anyDirty && form.$anyInvalid) {
+        return "Invalid form";
+      }
+      if (!form.$anyDirty) {
+        return "Please populate the form";
+      }
+      if (form.$errors.length > 0) {
+        console.log(form.$errors);
+        return "Error";
+      }
+
+      return "Submit";
+    });
+
+    const onSubmit = e => {
+      e.preventDefault();
+      if (form.$anyInvalid) {
+        alert("invalid form");
+      } else {
+        alert("submit form");
+      }
     };
 
     return {
       onSubmit,
+      submitText,
       form
     };
   }
@@ -68,11 +99,17 @@ export default createComponent({
 
 <style scoped>
 .invalid {
+  color: #aa2233;
+  background: grey;
 }
 
 .dirty {
+  color: yellow;
+  background: grey;
 }
 
-.valid {
+.error {
+  color: red;
+  background: grey;
 }
 </style>